Move DateTime classes to a directory
