+++ src/DataService/Access/Account.php
@@ -4,7 +4,7 @@ namespace Lullabot\Mpx\DataService\Access;
 
 use GuzzleHttp\Psr7\Uri;
 use Lullabot\Mpx\DataService\Annotation\DataService;
-use Lullabot\Mpx\DataService\NullDateTime;
+use Lullabot\Mpx\DataService\DateTime\NullDateTime;
 use Lullabot\Mpx\DataService\ObjectBase;
 use Lullabot\Mpx\DataService\PublicIdentifierInterface;
 use Lullabot\Mpx\DataService\IdInterface;
@@ -22,7 +22,7 @@ class Account extends ObjectBase implements PublicIdentifierInterface, IdInterfa
     /**
      * The date and time that this object was created.
      *
-     * @var \Lullabot\Mpx\DataService\DateTimeFormatInterface
+     * @var \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface
      */
     protected $added;
 
@@ -116,7 +116,7 @@ class Account extends ObjectBase implements PublicIdentifierInterface, IdInterfa
     /**
      * The date and time this object was last modified. The data service populates this field with a current timestamp each time the object is updated. See Retrieving Account objects for information about retrieving Account objects based on this field's value.
      *
-     * @var \Lullabot\Mpx\DataService\DateTimeFormatInterface
+     * @var \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface
      */
     protected $updated;
 
@@ -137,9 +137,9 @@ class Account extends ObjectBase implements PublicIdentifierInterface, IdInterfa
     /**
      * Returns the date and time that this object was created.
      *
-     * @return \Lullabot\Mpx\DataService\DateTimeFormatInterface
+     * @return \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface
      */
-    public function getAdded(): \Lullabot\Mpx\DataService\DateTimeFormatInterface
+    public function getAdded(): \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface
     {
         if (!$this->added) {
             return new NullDateTime();
@@ -151,9 +151,9 @@ class Account extends ObjectBase implements PublicIdentifierInterface, IdInterfa
     /**
      * Set the date and time that this object was created.
      *
-     * @param \Lullabot\Mpx\DataService\DateTimeFormatInterface $added
+     * @param \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface $added
      */
-    public function setAdded(\Lullabot\Mpx\DataService\DateTimeFormatInterface $added)
+    public function setAdded(\Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface $added)
     {
         $this->added = $added;
     }
@@ -419,9 +419,9 @@ class Account extends ObjectBase implements PublicIdentifierInterface, IdInterfa
     /**
      * Returns the date and time this object was last modified. The data service populates this field with a current timestamp each time the object is updated. See Retrieving Account objects for information about retrieving Account objects based on this field's value.
      *
-     * @return \Lullabot\Mpx\DataService\DateTimeFormatInterface
+     * @return \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface
      */
-    public function getUpdated(): \Lullabot\Mpx\DataService\DateTimeFormatInterface
+    public function getUpdated(): \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface
     {
         if (!$this->updated) {
             return new NullDateTime();
@@ -433,9 +433,9 @@ class Account extends ObjectBase implements PublicIdentifierInterface, IdInterfa
     /**
      * Set the date and time this object was last modified. The data service populates this field with a current timestamp each time the object is updated. See Retrieving Account objects for information about retrieving Account objects based on this field's value.
      *
-     * @param \Lullabot\Mpx\DataService\DateTimeFormatInterface $updated
+     * @param \Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface $updated
      */
-    public function setUpdated(\Lullabot\Mpx\DataService\DateTimeFormatInterface $updated)
+    public function setUpdated(\Lullabot\Mpx\DataService\DateTime\DateTimeFormatInterface $updated)
     {
         $this->updated = $updated;
     }
