+++ tests/src/Unit/DataService/FieldTest.php
@@ -2,7 +2,7 @@
 
 namespace Lullabot\Mpx\Tests\Unit\DataService;
 
-use Lullabot\Mpx\DataService\ConcreteDateTime;
+use Lullabot\Mpx\DataService\DateTime\ConcreteDateTime;
 use Lullabot\Mpx\DataService\DataServiceExtractor;
 use Lullabot\Mpx\DataService\Field;
 
@@ -47,7 +47,7 @@ class FieldTest extends ObjectTestBase
     public function getSetMethods()
     {
         $tests = parent::getSetMethods();
-        $tests['added'] = ['added', new ConcreteDateTime(\DateTime::createFromFormat('U.u', '1236030615.000'))];
+        $tests['added'] = ['added', new \Lullabot\Mpx\DataService\DateTime\ConcreteDateTime(\DateTime::createFromFormat('U.u', '1236030615.000'))];
         $tests['updated'] = ['updated', new ConcreteDateTime(\DateTime::createFromFormat('U.u', '1236030615.000'))];
 
         return $tests;
